+++ gnomes/gnomon.go
@@ -335,7 +335,7 @@ func CheckOwner(scid string) bool {
 
 	owner, _ := gnomes.GetSCIDValuesByKey(scid, "owner")
 	if owner != nil {
-		return owner[0] == rpc.Wallet.Address
+		return rpc.Wallet.IsAddress(owner[0])
 	}
 
 	return false